Move peak finding tests: pixstem test_dask_tools to test_peak_finding
Also change them to work for the newly added find_peak_dask functions.
